Fix integration/openconfig-interfaces tests
Ping the version of openconfig/public to 8527e6426cf1, which is the
most recent change to the relevant YANG model before the last update
to expected output in the pyangbind tests.

Also, patch two YANG models that trigger a pyang error (the same
change has since been done in openconfig/public).